Use View Model for Post Show. Extract pretty timesheet logic from model to controller
