+++ Dottle/Models/PostModel.cs
@@ -23,29 +23,9 @@ namespace Dottle.Models
         public string Email { get; set; }
         [Required(ErrorMessage = "Enter business address")]
         public string Address { get; set; }
+
         [Required(ErrorMessage = "Select working hours")]
         public string TimeSheet { get; set; }
 
-        public string PrettyWorkingHours()
-        {
-            StringBuilder sb = new StringBuilder("");
-            var timeSheet = JsonConvert.DeserializeObject<List<WorkingDay>>(this.TimeSheet);
-            foreach (WorkingDay day in timeSheet)
-            {
-                if (string.IsNullOrEmpty(day.DayName)) continue;
-                sb.Append("<div>");
-                sb.Append(day.DayName + " - ");
-                sb.Append("from: " + NumberToTime(day.HourFrom) + ":" + NumberToTime(day.MinuteFrom) + " ");
-                sb.Append("to: " + NumberToTime(day.HourTo) + ":" + NumberToTime(day.MinuteTo) + "\n");
-                sb.Append("</div>");
-            }
-
-            return sb.ToString();
-        }
-        private string NumberToTime (string time)
-        {
-            if (System.Int32.Parse(time) < 10) time = "0" + time;
-            return time;
-        }
     }
 }
