+++ Dottle/Models/PostShowViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Dottle.Models
+{
+    public class PostShowViewModel
+    {
+        public string PrettyTimeSheet;
+        public PostModel Post;
+    }
+}
