+++ Dottle/Views/Posts/Show.cshtml
@@ -1,30 +1,30 @@
-﻿@model PostModel
+﻿@model PostShowViewModel
 
 <div class="container">
     <div class="card">
         <div class="card-header">
-            <h3 class="card-title">@Model.Title</h3>
-            <h6 class="card-subtitle mb-2 text-muted">@Model.Description</h6>
+            <h3 class="card-title">@Model.Post.Title</h3>
+            <h6 class="card-subtitle mb-2 text-muted">@Model.Post.Description</h6>
         </div>
         <div class="card-body">
             <ul class="list-group list-grou-flush">
                 <h6 class="card-title">Contacts</h6>
                 <li class="list-group-item">
                     <div>
-                        <b>Phone number:</b> @Model.PhoneNumber
+                        <b>Phone number:</b> @Model.Post.PhoneNumber
                     </div>
                     <div>
-                        <b>Email:</b> @Model.Email
+                        <b>Email:</b> @Model.Post.Email
                     </div>
                     <div>
-                        <b>Address:</b> @Model.Address
+                        <b>Address:</b> @Model.Post.Address
                     </div>
                 </li>
                 
                 <h6 class="card-title" style="margin-top: 8px;">Working hours</h6>
                 <li class="list-group-item">
                     <div>
-                        @Html.Raw(@Model.PrettyWorkingHours())
+                        @Html.Raw(@Model.PrettyTimeSheet)
                     </div>
                 </li>
             </ul>
